fix: adjust error page in units
