+++ src/config/publicPaths.ts
@@ -1,0 +1,1 @@
+export const publicPaths = ["/error"];
